GPT Chatbot/chatbot/requirements.txt
django==5.0.2
djangorestframework==3.15.2
django-cors-headers==3.8.0
psycopg2==2.9.2
python-dotenv==0.19.0
openai==1.65.4
mysqlclient==2.1.0
gunicorn==20.1.0